Refactor / cleanup code
