+++ src/ApplicationFactory.ts
@@ -4,12 +4,12 @@ import { InversifyExpressServer } from 'inversify-express-utils';
 
 import { createContainer } from 'app/ContainerFactory';
 import { defaultHandler, notFoundHandler } from 'app/middleware/ErrorHandler';
-import { getExpressAppConfig } from 'app/utils/ConfigLoader';
+import { createExpressConfigFunction } from 'app/utils/ExpressConfigFunctionFactory';
 
 export class ApplicationFactory {
     public static createInstance(container: Container = createContainer()): Application {
         const server = new InversifyExpressServer(container);
-        server.setConfig(getExpressAppConfig(__dirname));
+        server.setConfig(createExpressConfigFunction(__dirname));
 
         const application: Application = server.build();
         application.use(notFoundHandler, defaultHandler);
