+++ src/Constants.ts
@@ -1,0 +1,1 @@
+export const APP_NAME = 'lfp-appeals-frontend';
